Add(kjh) : 채팅기능 + 코드 리팩토링 #4
