+++ src/main/java/com/example/livingbymyselfserver/common/entity/Like.java
@@ -1,4 +1,4 @@
-package com.example.livingbymyselfserver.like.entity;
+package com.example.livingbymyselfserver.common.entity;
 
 import com.example.livingbymyselfserver.common.entity.TimeStamped;
 import com.example.livingbymyselfserver.user.User;
