+++ src/main/java/com/example/livingbymyselfserver/community/Community.java
@@ -3,10 +3,9 @@ package com.example.livingbymyselfserver.community;
 import com.example.livingbymyselfserver.comment.entity.CommunityComment;
 import com.example.livingbymyselfserver.common.entity.TimeStamped;
 import com.example.livingbymyselfserver.community.dto.CommunityRequestDto;
-import com.example.livingbymyselfserver.like.entity.CommunityLike;
+import com.example.livingbymyselfserver.community.like.entity.CommunityLike;
 import com.example.livingbymyselfserver.user.User;
 import jakarta.persistence.*;
-import java.time.Instant;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.DynamicUpdate;
