+++ src/main/java/com/example/livingbymyselfserver/groupBuying/GroupBuyingServiceImpl.java
@@ -2,14 +2,12 @@ package com.example.livingbymyselfserver.groupBuying;
 
 import com.example.livingbymyselfserver.attachment.S3Service;
 import com.example.livingbymyselfserver.attachment.entity.Attachment;
-import com.example.livingbymyselfserver.attachment.entity.AttachmentCommunityUrl;
 import com.example.livingbymyselfserver.attachment.entity.AttachmentGroupBuyingUrl;
 import com.example.livingbymyselfserver.attachment.fair.AttachmentGroupBuyingUrlRepository;
 import com.example.livingbymyselfserver.common.ApiResponseDto;
 import com.example.livingbymyselfserver.common.PostTypeEnum;
 import com.example.livingbymyselfserver.common.RedisUtil;
 import com.example.livingbymyselfserver.common.RedisViewCountUtil;
-import com.example.livingbymyselfserver.community.dto.CommunityResponseDto;
 import com.example.livingbymyselfserver.groupBuying.application.ApplicationUsers;
 import com.example.livingbymyselfserver.groupBuying.application.ApplicationUsersRepository;
 import com.example.livingbymyselfserver.groupBuying.dto.GroupBuyingDetailResponseDto;
@@ -25,10 +23,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -133,54 +129,16 @@ public class GroupBuyingServiceImpl implements GroupBuyingService {
     }
 
     double viewCount = redisViewCountUtil.getViewPostCount(groupBuyingId.toString(),PostTypeEnum.GROUPBUYING) ==null?1:redisViewCountUtil.getViewPostCount(groupBuyingId.toString(),PostTypeEnum.GROUPBUYING);
+    List<User> users = applicationUsersRepository.findAllByGroupBuyingId(groupBuyingId);
 
     if (attachmentGroupBuyingUrl == null) {
-      return new GroupBuyingDetailResponseDto(groupBuying,viewCount);
+      return new GroupBuyingDetailResponseDto(groupBuying,viewCount, users);
     } else {
-      return new GroupBuyingDetailResponseDto(groupBuying,attachmentGroupBuyingUrl, viewCount);
+      return new GroupBuyingDetailResponseDto(groupBuying,attachmentGroupBuyingUrl, viewCount, users);
     }
   }
 
-  @Override
-  @Transactional
-  public ApiResponseDto createApplication(User user, Long groupBuyingId) {
 
-    GroupBuying groupBuying = findGroupBuying(groupBuyingId);
-    if (applicationUsersRepository.existsByGroupBuyingAndUser(groupBuying, user))
-      throw new IllegalArgumentException("이미 신청한 공고입니다.");
-    if (groupBuying.getEnumStatus() == GroupBuyingStatusEnum.DEADLINE)
-      throw new IllegalArgumentException("마감된 공고입니다.");
-    //유저에 캐쉬 추가한 뒤 캐쉬 확인 차감작업
-    if (user.getCash() < groupBuying.getPerUserPrice()) {
-      throw new IllegalArgumentException("캐쉬가 부족합니다. 충전해주세요!");
-    }
-    if (groupBuying.getAppUsers().size() + 1 > groupBuying.getMaxUser()) {
-      throw new IllegalArgumentException("인원이 가득 찼습니다.");
-    }
-    ApplicationUsers applicationUsers = new ApplicationUsers(user, groupBuying);
-    user.setCash(user.getCash()-groupBuying.getPerUserPrice());
-    applicationUsersRepository.save(applicationUsers);
-    return new ApiResponseDto("공동구매 신청완료", 200);
-    }
-
-
-
-  @Override
-  @Transactional
-  public ApiResponseDto deleteApplication(User user, Long groupBuyingId) {
-    GroupBuying groupBuying = findGroupBuying(groupBuyingId);
-
-    if (!applicationUsersRepository.existsByGroupBuyingAndUser(groupBuying, user)) {
-      throw new IllegalArgumentException("해당 공고를 신청한 유저가 아닙니다.");
-    }
-    redisUtil.delete("GroupBuying:"+user.getId()+groupBuyingId);
-    //삭제시 redis내 같은 키값 조회수 데이터도 삭제해주기
-    ApplicationUsers applicationUsers = applicationUsersRepository.findByGroupBuyingAndUser(groupBuying, user);
-    user.setCash(user.getCash()+groupBuying.getPerUserPrice());
-    applicationUsersRepository.delete(applicationUsers);
-
-    return new ApiResponseDto("공고 지원 취소", 200);
-  }
 
   @Override
   public List<GroupBuyingResponseDto> getGroupBuyingList(User user, Pageable pageable) {
