+++ src/main/java/com/example/livingbymyselfserver/groupBuying/application/ApplicationUsersRepository.java
@@ -2,6 +2,7 @@ package com.example.livingbymyselfserver.groupBuying.application;
 
 import com.example.livingbymyselfserver.groupBuying.GroupBuying;
 import com.example.livingbymyselfserver.user.User;
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,4 +11,6 @@ public interface ApplicationUsersRepository extends JpaRepository<ApplicationUse
     Boolean existsByGroupBuyingAndUser(GroupBuying groupBuying, User user);
 
     ApplicationUsers findByGroupBuyingAndUser(GroupBuying groupBuying, User user);
+
+    List<User> findAllByGroupBuyingId(Long groupBuying_id);
 }
