+++ src/main/java/com/example/livingbymyselfserver/groupBuying/application/ApplicationUsersService.java
@@ -1,0 +1,11 @@
+package com.example.livingbymyselfserver.groupBuying.application;
+
+import com.example.livingbymyselfserver.common.ApiResponseDto;
+import com.example.livingbymyselfserver.user.User;
+
+
+public interface ApplicationUsersService {
+  ApiResponseDto createApplication(User user, Long groupBuyingId);
+
+  ApiResponseDto deleteApplication(User user, Long groupBuyingId);
+}
