+++ src/main/java/com/example/livingbymyselfserver/user/User.java
@@ -18,7 +18,7 @@ import org.hibernate.annotations.DynamicUpdate;
 
 @Entity
 @Getter
-@Table(name = "user")
+@Table(name = "users")
 @NoArgsConstructor
 @DynamicUpdate
 public class User {
