+++ src/main/java/com/example/livingbymyselfserver/user/dto/SignupRequestDto.java
@@ -8,4 +8,5 @@ public class SignupRequestDto {
   private String password;
   private String passwordCheck;
   private String phoneNumber;
+  private String adress;
 }
